refactor(cli): move the validation layer to lib
The layer carries contextual logic rather than generic helpers: it filters
issues by level, groups them by entity, formats pretty and JSON output, and
writes the run summary to the reporter. That puts it alongside `ui`, `config`,
`logger`, and `reporter` as a core module, not in `utils`.

Pure move plus import paths. Documented in the CLI's directory table.
