+++ packages/cli/src/lib/validation/format-json.ts
