+++ packages/cli/src/lib/validation/format-pretty.ts
