+++ packages/cli/src/lib/validation/index.ts
